+++ include/vsmc/rng/uniform_real_distribution.hpp
@@ -76,7 +76,7 @@ inline void uniform_real_distribution_impl_u64(RNGType &rng, std::size_t n,
     if (n % 2 != 0)
         u[n - 1] = static_cast<std::uint32_t>(rng());
     for (std::size_t i = 0; i != n; ++i)
-        r[i] = U01<std::uint64_t, RealType, Closed, Open>::eval(u[i]);
+        r[i] = U01<std::uint32_t, RealType, Closed, Open>::eval(u[i]);
     RealType d = b - a;
     for (std::size_t i = 0; i != n; ++i)
         r[i] = a + d * r[i];
